wallet_base/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
black==25.1.0
celery==5.4.0
cffi==1.17.1
cfgv==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==44.0.0
distlib==0.3.9
Django==4.2.19
django-aesfield==4.0.0
django-celery-beat==2.7.0
django-redis==5.4.0
django-timezone-field==7.1
djangorestframework==3.15.2
filelock==3.17.0
gunicorn==23.0.0
identify==2.6.6
iniconfig==2.0.0
kombu==5.4.2
m2secret-py3==1.3
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycparser==2.22
pytest==8.3.4
pytest-django==4.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
raven==6.10.0
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
vine==5.1.0
virtualenv==20.29.1
wcwidth==0.2.13
